compute coverage at intervals
Compute coverage if file was found after N ms from the previous time
coverage was computed. This is mostly useful for post-mortem analysis to
speed up coverage gathering and still have a nice time resolution for
plots and whatnot.

This works by parsing the time a test case was found from the test
filename. AFL++ already operates like this by default.

Also adds the ability to store coverage as CSV.